# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Rest",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Fetch Students list",
                "description": "Fetch Students list",
                "operationId": "66992b07a24419f015b23b650c8da5b1",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "id": "9d12795b-138b-47c9-912e-01eff34cdaa5",
                                            "name": "Student",
                                            "email": "student@example.com",
                                            "active": true,
                                            "created_at": "2024-09-22T17:38:29.000000Z",
                                            "updated_at": "2024-09-22T17:38:29.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/students/{uuid}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Fetch Student",
                "description": "Fetch Student",
                "operationId": "2b0cd8250bdcd5f0ff97c959fd970614",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": "9d12795b-138b-47c9-912e-01eff34cdaa5",
                                        "name": "Student",
                                        "email": "student@example.com",
                                        "active": true,
                                        "created_at": "2024-09-22T17:38:29.000000Z",
                                        "updated_at": "2024-09-22T17:38:29.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Update Student",
                "description": "Update Student",
                "operationId": "c20624e1d51d8197ac6b251ced58760e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "current_password": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": "9d12795b-138b-47c9-912e-01eff34cdaa5",
                                        "name": "Student",
                                        "email": "student@example.com",
                                        "active": true,
                                        "created_at": "2024-09-22T17:38:29.000000Z",
                                        "updated_at": "2024-09-22T17:38:29.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create Student",
                "description": "Create Student",
                "operationId": "64fe5598b19d6227a380863a761989bc",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": "9d12795b-138b-47c9-912e-01eff34cdaa5",
                                        "name": "Student",
                                        "email": "student@example.com",
                                        "active": true,
                                        "created_at": "2024-09-22T17:38:29.000000Z",
                                        "updated_at": "2024-09-22T17:38:29.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Delete Student",
                "description": "Delete Student",
                "operationId": "dd828443b1721410b49e1401cda05e7b",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
...
}